BBG-0002: refactor
