+++ src/main/java/al/tonikolaba/entity/enemies/RedEnergy.java
@@ -1,6 +1,5 @@
 package al.tonikolaba.entity.enemies;
 
-import al.tonikolaba.entity.Enemy;
 import al.tonikolaba.handlers.Content;
 import al.tonikolaba.tilemap.TileMap;
