+++ src/main/java/al/tonikolaba/gamestate/Level3State.java
@@ -4,7 +4,7 @@ All rights reserved ©.
 
 package al.tonikolaba.gamestate;
 
-import al.tonikolaba.entity.Enemy.EnemyType;
+import al.tonikolaba.entity.enemies.Enemy.EnemyType;
 import al.tonikolaba.entity.Player;
 import al.tonikolaba.tilemap.Background;
